start working on tokenisation
