+++ tokenisation/fileCleaner.h
@@ -1,4 +1,7 @@
-#include "../NotWorking/header.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <ctype.h>
 
 void removeCommentsAndEmptyLines(char *str) {
 
@@ -102,41 +105,4 @@ void tokenizeLine(char *line, char cleanedLines[][3][1000], int *numLines) {
         // Increment numLines even if there's only one token
         (*numLines)++;
     }
-}
-
-int main(int argc, char *argv[]) {
-
-    // Check if a file name was passed as an argument to the program
-    if (argc != 2) {
-        printf("Error: Please pass in a file name\n");
-        return 1;
-    }
-
-    FILE *file = fopen(argv[1], "r");
-    if (file == NULL) {
-        printf("Error: File does not exist\n");
-        return 1;
-    }
-
-    char line[1000];
-    char cleanedLines[100][3][1000];
-    int numLines = 0;
-
-    // Read the file line by line, clean each line, and tokenize each line
-    while (fgets(line, sizeof(line), file)) {
-        removeCommentsAndEmptyLines(line);
-        tokenizeLine(line, cleanedLines, &numLines);
-    }
-
-    fclose(file);
-
-    // Print out the cleaned lines
-    for (int i = 0; i < numLines; i++) {
-        printf("{\"%s\", \"%s\", \"%s\"}\n",
-                cleanedLines[i][0],
-                cleanedLines[i][1],
-                cleanedLines[i][2]);
-    }
-
-    return 0;
 }
